requirements.txt
Flask==1.1.1
Flask-Cors==3.0.10
xrpl-py==1.1.1
cryptography==3.0
requests==2.24.0